OvmfPkg: Add SVSM X64 package and FW config
Adds a new OVMF package that bundles a pre-built SVSM
module as part of the firmware volume. The SVSM
module binary is located by default in Svsm/svsm.bin
 but this can be overridden by defining SVSM_MODULE_FILE
at build time.

Currently supports only a 4MB FV with the space
for the SVSM module being provided by reducing the
size of the FVMAIN_COMPACT volume.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>