OvmfPkg/ResetVector: Support launch of SVSM module
For builds using the OvmfPkgSvsmX64 package, the reset
vector launches a binary SVSM module in the FV instead of
launching OVMF directly.

The  metadata in the top 4K of the FV relates to the SVSM
module and not the OVMF firmware. The SVSM module will
look for the OVMF metadata 4K below this before launching
OVMF at the fixed address of 0xffffefe0.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>